rust: Refactor LabelDefinition to Iterator
The public API exposed the LabelDefinition struct and expected users
to initialize a list of LabelDefinition structs, which are much more
naturally represented by some Map from &str to u64.

This commit refactors the rust assemble functions to take an iterator of
`(&&str, &u64)` instead of `&[LabelDefinition]`. The iterator type was
chosen since both `&HashMap` and `&BTreeMap` use an iterator of type
`(&K, &V)`, allowing the user to decide on the exact Map implementation
they want to use.