Fix resolving of `adrp`
The `adrp` instruction was wrongly resolved by us. We computed the
immediate to be `alignUp(offset, 4096)`, where offset was the difference
between the current location and the label. In actuality, we needed
to compute `imm = page(label) - page(current_addr)`. This commit
uses the provided start address to compute the absolute addresses
of the current instruction, using it to compute the correct immediate.