Ensures an error for unhandled relocations
Currently, when encountering a relocation, Nyxstone uses the wrapped
ELFObjectWriter to record the relocation. It turns out that this means
that LLVM will simply emit an immediate of zero for the instruction,
since a relocation would normally be handled by the linker. This leads
to invalid instruction bytes for such instructions. Additionally, for
instructions where the label is not defined, LLVM also emits a
relocation for the linker, since it can not infer the location of the
label.

This commit ensures that we emit an error for any relocation which we do
not handle ourselfes, ensuring that
- Undefined labels are reported
- Instructions which can not be assembled are reported as an error