Refactors the rust generics for label definitions (#53)
Previously, the generics for label definitions enforced that the passed
container is `IntoIterator<Item = (&&str, &u64)>`. As a consequence,
the container was unable to own the label names and could only hold a
`&str`. When implementing the rust cli example, we noticed that this
really impractical, since it forces the caller to hold two containers,
one which owns the label names as `String`s and one which can be passed
to the functions, holding the `&str` references to the first container.

This commit tries to make the generics more practical by only requiring
the label name to implement `AsRef<str>`. This change allows the passed
container to own the label names as `String`s.

One downside of this change is that whenever no labels need to be
passed, the caller needs to specify the exact types in the container,
like this:
```rust
nyxstone.assemble("xor rax,rax", 0, &HashMap::<&str, u64>::new())
```
To keep the API ergonomic, this commit moves the assemble functions
to `assemble_with` functions, and replaces the original functions so
that they no longer take a map of labels. This way the caller can simply
call `assemble` if they do not need to provide labels instead of
specifying the exact type of the map container every time.