Fix adrp relocation and emit error on unhandled relocation (includes missing label) (#54)
* Fix resolving of `adrp`

The `adrp` instruction was wrongly resolved by us. We computed the
immediate to be `alignUp(offset, 4096)`, where offset was the difference
between the current location and the label. In actuality, we needed
to compute `imm = page(label) - page(current_addr)`. This commit
uses the provided start address to compute the absolute addresses
of the current instruction, using it to compute the correct immediate.

* Ensures an error for unhandled relocations

Currently, when encountering a relocation, Nyxstone uses the wrapped
ELFObjectWriter to record the relocation. It turns out that this means
that LLVM will simply emit an immediate of zero for the instruction,
since a relocation would normally be handled by the linker. This leads
to invalid instruction bytes for such instructions. Additionally, for
instructions where the label is not defined, LLVM also emits a
relocation for the linker, since it can not infer the location of the
label.

This commit ensures that we emit an error for any relocation which we do
not handle ourselfes, ensuring that
- Undefined labels are reported
- Instructions which can not be assembled are reported as an error

* Fix formatting

* Fix adrp test

Fixes the test for `adrp` where the start address is `0x1000` and
the label is at `0x1010`. Since the address and the label are on the
same 4096 byte aligned page, the correct immediate for `adrp` is zero,
and not 4096, which we incorrectly put in the test.

* Fix cppcheck