Add recommended triples to docs
Currently, we do not provide any examples for triples and use shorthand
triples in all examples. This gives way to the assumption that all
triples (and their shorthands) are created equal, when in reality they
are not (see #63 for reference).

This commit documents common triples we use internally and makes it
clearer that triple shorthands can decrease Nyxstones ability to asseble
certain instructions.