Use cmake_package to link against LLVM
Currently, we are using the same approach as the llvm-sys package to
facilitate linking against LLVM, replicating a lot of their 'build.rs'
in the process. The process uses the `llvm-config` binary to find the
necessary libraries and include directories to link against LLVM.

This method has some downsides, mainly not being able to find the binary
if it is not in the PATH and `llvm-config` not properly reporting that
FFI needs to be linked. We noticed that these problems do not occur for
the CMake build of the C++ Nyxstone library. This commit introduces the
`cmake_package` crate to use CMake's `find_library` function to find the
LLVM libraries.